Move RFIDLinkRepository to service package.
