+++ src/main/java/ch/wisv/areafiftylan/extras/rfid/service/RFIDServiceImpl.java
@@ -19,7 +19,6 @@ package ch.wisv.areafiftylan.extras.rfid.service;
 
 import ch.wisv.areafiftylan.exception.*;
 import ch.wisv.areafiftylan.extras.rfid.model.RFIDLink;
-import ch.wisv.areafiftylan.extras.rfid.model.RFIDLinkRepository;
 import ch.wisv.areafiftylan.products.model.Ticket;
 import ch.wisv.areafiftylan.products.service.TicketService;
 import ch.wisv.areafiftylan.users.model.User;
